Carlson Past Program
The second part of the project is to use database programming with JSP
web pages.

Requirements: There is an existing database file and source code that
can be used to load / save the database relations (tables) into / from
arrays. Thus, the student is not required to come up with any new
database tables but he/she has the choice to make any updates if needed.
The database source code is generated by MySQL server. The Java source
code is an Eclipse project. Please refer to how to setup a database on
the previous chapter.

Authentication:

The system should be able to identify the user based on his type
(student, staff or faculty) and load the appropriate webpage that is
related to him/her. A user can login/logout any time from the system. A
sample authentication source code has been provided during JSP
programming chapter, it can be used for this project. To make it
simpler, the user can login using his email (no need for password). (2
pts)

The student is able to:

Check the classes’ descriptions (timing, instructor, etc.) that he/she
is enrolled in. (1 pt)
Change some of his own account information (address and phone). (2 pt)

The faculty is able to:

Check the classes’ descriptions (timing, instructor, etc.) that he/she
is teaching. (1 pt)
Change some of his own account information (address, office address,
office hours and phone). (2 pt)

The staff is able to:

Add students to the system (2 pt)
Add more faculty to the system (2 pt)
Add courses to the system. (2 pts)
Modify course information. (2 pts)